new fix

+++ app/src/main/java/com/xianggao/smartbutler/entity/SensorData.java
@@ -11,9 +11,9 @@ package com.xianggao.smartbutler.entity;
 
 public class SensorData {
     private int type;
-    private float x;
-    private float y;
-    private float z;
+    private double x;
+    private double y;
+    private double z;
     private long timeline;
 
     public int getType() {
@@ -24,27 +24,27 @@ public class SensorData {
         this.type = type;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(float z) {
+    public void setZ(double z) {
         this.z = z;
     }
